revert: restart v0.7.0 with release automation

+++ src/output/ringPolymerEnergyOutput.cpp
@@ -24,10 +24,10 @@
 
 #include <algorithm>   // for for_each
 #include <format>      // for format
-#include <numeric>
-#include <vector>   // for vector
+#include <vector>      // for vector
 
 #include "physicalData.hpp"   // for PhysicalData
+#include "stlVector.hpp"      // for sum
 
 using namespace output;
 using namespace physicalData;
@@ -57,8 +57,7 @@ void RingPolymerEnergyOutput::write(
 
     _fp << std::format(
         "{:20.12f}\t",
-        sumOfRingPolymerEnergies(dataVector) /
-            static_cast<double>(dataVector.size())
+        sumOfRingPolymerEnergies(dataVector) / dataVector.size()
     );
 
     std::ranges::for_each(
